python: refactor common routine into binding_common.py
Some of the routines will be accessed by multiple binding scripts and
they have somewhat common context. Move them to binding_common.py for
better access.
